Исправление update и delete

+++ api/crud.go
@@ -88,7 +88,12 @@ func DeleteUser(c echo.Context) error {
 		return c.JSON(http.StatusBadRequest,
 			map[string]string{"error": "Ошибка ID формата"})
 	}
-
+	currentUserID := c.Get("user_id").(int)
+	if id == currentUserID {
+		return c.JSON(http.StatusBadRequest, map[string]string{
+			"error": "Нельзя удалить самого себя",
+		})
+	}
 	// Используем интерфейс UserService
 	err = userService.DeleteUser(ctx, id)
 	if err != nil {
@@ -96,12 +101,6 @@ func DeleteUser(c echo.Context) error {
 			return c.JSON(http.StatusNotFound, map[string]string{"error": "Пользователь не найден"})
 		}
 		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
-	}
-	currentUserID := c.Get("user_id").(int)
-	if id == currentUserID {
-		return c.JSON(http.StatusBadRequest, map[string]string{
-			"error": "Нельзя удалить самого себя",
-		})
 	}
 
 	return c.JSON(http.StatusOK, map[string]string{
